+++ uk-manager/src/bnp/effects.rs
@@ -8,6 +8,7 @@ impl BnpConverter {
     pub fn handle_effects(&self) -> Result<()> {
         let effects_path = self.path.join("logs/effects.yml");
         if effects_path.exists() {
+            log::debug!("Processing status effect log");
             let mut base = Byml::from_binary(
                 self.dump
                     .get_bytes_from_sarc("Pack/Bootup.pack//Ecosystem/StatusEffectList.sbyml")?,