+++ uk-manager/src/bnp/events.rs
@@ -13,6 +13,7 @@ impl BnpConverter {
     pub fn handle_events(&self) -> Result<()> {
         let events_path = self.path.join("logs/eventinfo.yml");
         if events_path.exists() {
+            log::debug!("Processing eventinfo log");
             let diff = EventInfo::from_byml(&Byml::from_text(fs::read_to_string(events_path)?)?)?;
             let base = self.dump.get_from_sarc(
                 "Events/EventInfo.product.byml",