+++ uk-manager/src/bnp/quests.rs
@@ -10,6 +10,7 @@ impl BnpConverter {
     pub fn handle_quests(&self) -> Result<()> {
         let quests_path = self.path.join("logs/quests.yml");
         if quests_path.exists() {
+            log::debug!("Processing quests log");
             let mut diff = Byml::from_text(fs::read_to_string(quests_path)?)?.into_hash()?;
             let mut quests = Byml::from_binary(
                 self.dump