+++ uk-manager/src/bnp/savedata.rs
@@ -9,6 +9,7 @@ impl BnpConverter {
     pub fn handle_savedata(&self) -> Result<()> {
         let save_path = self.path.join("logs/savedata.yml");
         if save_path.exists() {
+            log::debug!("Processing savedata log");
             let mut diff = Byml::from_text(fs::read_to_string(save_path)?)?.into_hash()?;
             let base = self.dump.get_from_sarc(
                 "GameData/savedataformat.sarc",